[bindings/python] Split exposeUnaryFunctions() out to its own .cpp file
